requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
idna==3.4
Pillow==9.2.0
pytz==2022.4
requests==2.28.1
sqlparse==0.4.2
tzdata==2022.2
urllib3==1.26.12
